feat: Completed rectangle module refactoring
