+++ CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.0)
 
 project(CPPMensuration)
-add_executable(main src/main.cpp src/display/display.cpp src/shapes/rectangle/rectangle.cpp)
+add_executable(main src/main.cpp src/display/display.cpp src/shapes/rectangle/rectangle.cpp src/utils/utils.cpp)
 set(CMAKE_CXX_STANDARD 11)
