+++ src/display/display.cpp
@@ -10,5 +10,5 @@ void displayTitle() {
     cout <<"\nThe available shapes are:" << endl;
     cout << " \n\t1. Rectangle \n\t2. Parallelogram \n\t3. Triangle \n\t4. Trapezium \n\t5. Square " << endl;
 
-    cout <<"\n Enter the shape's number to continue: ";
+    cout <<"\nEnter the shape's number to continue: ";
 }
