+++ src/shapes/rectangle/rectangle.h
@@ -1,2 +1,4 @@
+void display();
+void handleShapeInput();
 void area();
 void perimeter();
