+++ src/utils/utils.h
@@ -1,0 +1,4 @@
+int getUserChoice();
+void lineBreak();
+void refreshStdOut();
+void handleShapeInput(int shape);
